% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganimate.R
\name{asciicast_file}
\alias{asciicast_file}
\title{Tag a character string as being for an asciicast file}
\usage{
asciicast_file(filename)
}
\arguments{
\item{filename}{asciicast filename. Should end in ".cast"}
}
\description{
Tag a character string as being for an asciicast file
}
