% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{asciicast_knitr_options}
\alias{asciicast_knitr_options}
\title{Default R options to set in the background R process for knits}
\usage{
asciicast_knitr_options()
}
\value{
List of options.
}
\description{
You can pass these options to \code{\link[=init_knitr_engine]{init_knitr_engine()}}, after possibly
overriding some of them.
}
\examples{
asciicast_knitr_options()
}
\seealso{
Other asciicast in Rmd: 
\code{\link{init_knitr_engine}()}
}
\concept{asciicast in Rmd}
