% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{asciicast_options}
\alias{asciicast_options}
\title{Default options to set in the asciicast subprocess.}
\usage{
asciicast_options()
}
\value{
Named list.
}
\description{
Default options to set in the asciicast subprocess.
}
\examples{
asciicast_options()
}
