% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganimate.R
\name{asciicast_renderer}
\alias{asciicast_renderer}
\title{A 'renderer' to use with \code{gganimate::animate()}}
\usage{
asciicast_renderer(filename = tempfile(fileext = ".cast"), ...)
}
\arguments{
\item{filename}{output file for the asciicast animation}

\item{...}{further arguments passed to \code{asciicast::create_asciicast_flipbook()}}
}
\value{
Returns a function which takes 2 arguments: \code{frames} and \code{fps}
}
\description{
Warning: this only currently works with a hacked, custom version of gganimate.
}
