% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{asciicast_start_process}
\alias{asciicast_start_process}
\title{Start an asciicast background process}
\usage{
asciicast_start_process(
  timeout = 10,
  allow_errors = TRUE,
  startup = NULL,
  record_env = NULL,
  echo = TRUE
)
}
\arguments{
\item{timeout}{Idle timeout, in seconds If the R subprocess running
the recording does not answer within this limit, it is killed and the
recording stops. Update this for slow running code, that produces no
output as it runs.}

\item{allow_errors}{Whether to cast errors properly. If this is set to
\code{TRUE}, then asciicast overwrites the \code{"error"} option. Only change
this if you know what you are doing.}

\item{startup}{Quoted language object to run in the subprocess before
starting the recording.}

\item{record_env}{Environment variables to set for the R subprocess.}

\item{echo}{Whether to echo the input to the terminal. If \code{FALSE}, then
only the output is shown.}
}
\description{
This is for expert use, if you want to run multiple recordings in the
same process.
}
\seealso{
Other asciicast functions: 
\code{\link{asciicast-package}},
\code{\link{default_theme}()},
\code{\link{play}()},
\code{\link{read_cast}()},
\code{\link{record}()},
\code{\link{write_json}()},
\code{\link{write_svg}()}
}
\concept{asciicast functions}
