% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidget.R
\name{asciinema}
\alias{asciinema}
\title{asciicast HTML widget}
\usage{
asciinema(cast, from = 0, height = NULL, width = NULL,
  autoplay = FALSE, loop = FALSE, speed = 1, title = NULL,
  author = NULL, author_url = NULL, author_img_url = NULL,
  poster_text = title, poster_frame = "", font_size = "small",
  theme = "asciinema", html_height = NULL, html_width = NULL,
  element_id = NULL)
}
\arguments{
\item{cast}{\code{asciicast} object.}

\item{from}{Where to start the playback from, in seconds.}

\item{height}{Number of rows, defaults to the number of rows in the
recording, or 24 if not specified in the cast.}

\item{width}{Number of columns, defaults to the number columns in the
recording, or 80 if not specified in the cast.}

\item{autoplay}{Whether to start playing the cast automatically.}

\item{loop}{Whether to loop the playback.}

\item{speed}{Whether to play slower or faster. 1 is normal speed.}

\item{title}{If specified, it overrides the title in the recording.}

\item{author}{Author, displayed in the titlebar in fullscreen mode.}

\item{author_url}{URL of the author's homepage/profile. Author name
(author above) is linked to this URL.}

\item{author_img_url}{URL of the author's image, displayed in the
titlebar in fullscreen mode.}

\item{poster_text}{if not \code{NULL}, used as the text of the poster
(preview).}

\item{poster_frame}{If not \code{NULL}, used as the image of the poster
(preview).}

\item{font_size}{Size of terminal font. Possible values: small, medium,
big, any css \code{font-size} value (e.g. 15px).}

\item{theme}{Theme.}

\item{html_height}{HTML height of the widget.}

\item{html_width}{HTML width of the widget.}

\item{element_id}{HTML id of the widget's element. If \code{NULL}, then the
id is generated randomly.}
}
\description{
asciicast HTML widget
}
