% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{capture_cast}
\alias{capture_cast}
\title{Capture Expression Output and Create a Cast Object}
\usage{
capture_cast(expr, num_colors = 256)
}
\arguments{
\item{expr}{An expression to evaluate and capture the output from}

\item{num_colors}{The number of colors to use in the terminal (default: 256)}
}
\value{
A cast object
}
\description{
This function captures the console output from evaluating an R expression
and returns a cast object.
}
\examples{
cast_obj <- capture_cast(tibble::tibble(a = 1:3))
}
