% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciicast.R
\name{create_asciicast_flipbook}
\alias{create_asciicast_flipbook}
\title{Create an asciinema flipbook from a list of txt files}
\usage{
create_asciicast_flipbook(txt_files, filename = "out.cast", fps = 20,
  width = NULL, height = NULL, verbose = FALSE, ...)
}
\arguments{
\item{txt_files}{character vector of txt filenames}

\item{filename}{output filename for asciicast animation}

\item{fps}{target frames-per-second of the animation. default: 20}

\item{width, height}{width and height of terminal. default: NULL determine
size as the maximum width and height of all the text files}

\item{verbose}{default: FALSE}

\item{...}{other arguments passed to \code{create_v2_header}}
}
\description{
This function simply joins together a list of text files into a single asciicast file.
Each of the input files becomes a frame in a very crude "flipbook" animation.
}
