% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciicast.R
\name{create_v2_header}
\alias{create_v2_header}
\title{Create a header for v2 asciicast}
\usage{
create_v2_header(width = NULL, height = NULL, title = NULL,
  env = NULL, ...)
}
\arguments{
\item{width, height}{dimensions of terminal}

\item{title}{title of asciicast}

\item{env}{terminal environment}

\item{...}{other arguments ignored}
}
\description{
Create a header for v2 asciicast
}
