% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{default_theme}
\alias{default_theme}
\title{The default asciicast theme}
\usage{
default_theme()
}
\value{
A named list.
}
\description{
Currently only used for \code{\link[=write_svg]{write_svg()}}
}
\seealso{
Other asciicast functions: \code{\link{asciicast-package}},
  \code{\link{asciicast_start_process}},
  \code{\link{play}}, \code{\link{read_cast}},
  \code{\link{record}}, \code{\link{write_json}},
  \code{\link{write_svg}}
}
\concept{asciicast functions}
