% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{get_locales}
\alias{get_locales}
\title{Helper function to query locales as a named character vector.}
\usage{
get_locales()
}
\value{
Named character vector with entries:
\itemize{
\item \code{LC_COLLATE}, \code{LC_CTYPE}, \code{LC_MONETARY}, \code{LC_NUMERIC} and \code{LC_TIME}.
}
}
\description{
Helper function to query locales as a named character vector.
}
