% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyrmd.R
\docType{data}
\name{lazyrmd}
\alias{lazyrmd}
\title{Configure the lazy vignette builder}
\format{
An object of class \code{standalone_lazyrmd} (inherits from \code{standalone}) of length 3.
}
\usage{
lazyrmd
}
\arguments{
\item{local}{Whether/when to build vignettes for local builds.}

\item{ci}{Whether/when to build vignettes on the CI(s).}

\item{cran}{Whether/when to build vignettes on CRAN.

Possible values for \code{local}, \code{ci} and \code{cran} are:
\itemize{
\item \code{TRUE}: always (re)build the vignettes.
\item \code{FALSE}: never (re)build the vignettes.
\item \code{"if-newer"}: Only re(build) the vignettes if the Rmd file is newer.
\item \code{"no-code"}: never rebuild the vignettes, and make sure that the
\code{.R} file only contains dummy code. (This is to make
sure that the \code{.R} file runs on CRAN.)
\item a function object. This is called without arguments and if it
returns an \code{isTRUE()} value, then the vignette is built.
Build vignettes
}

This is a utility function, and you would probably not use it in the
package code itself. It can force the re-building of all vignettes.
This function ignores the vignette configuration that was created with
\code{onload_hook()}.}

\item{mode}{\code{"force"} forces vignette re-building, even if the \code{.Rmd}
file is older than the output file(s). \code{"if-newer"} will only
re-build if the \code{.Rmd} file is newer.}
}
\value{
The return value of \code{\link[tools:buildVignettes]{tools::buildVignettes()}}.
This is called by \link[tools:buildVignettes]{tools::buildVignettes}, as the weave function of
the lazyrmd vignette engine
This is called by \link[tools:buildVignettes]{tools::buildVignettes}, as the tangle function of
the lazyrmd vignette engine
If NOT_CRAN is set to a true value, then we are local
Otherwise we are local if not running inside \verb{R CMD check}.
}
\description{
Configure the lazy vignette builder
}
\keyword{datasets}
