% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciicast.R, R/gganimate.R
\name{play_asciicast}
\alias{play_asciicast}
\alias{print.asciicast}
\title{Play an asciicast file}
\usage{
play_asciicast(x, autoplay = FALSE, loop = TRUE,
  theme = c("black-on-wite", "asciinema", "tango", "solarized-dark",
  "solarized-light", "monokai"), verbose = FALSE, ...)

\method{print}{asciicast}(x, autoplay = FALSE, loop = TRUE,
  theme = c("black-on-wite", "asciinema", "tango", "solarized-dark",
  "solarized-light", "monokai"), verbose = FALSE, ...)
}
\arguments{
\item{x}{asciicast filename}

\item{autoplay}{default: FALSE}

\item{loop}{default: FALSE}

\item{theme}{terminal colour theme. default: black-on-white}

\item{verbose}{default: FALSE}

\item{...}{further arguments passed to of from other methods.}
}
\description{
This function creates an HTML document (which includes the javascript and
CSS for the player).   It then asks Rstudio to show this HTML in the viewer, or
if that is not available, to open it in a browser.
}
\details{
See \url{https://github.com/asciinema/asciinema-player} for information on the player.
}
