% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{print_html_style}
\alias{print_html_style}
\title{CSS for a theme}
\usage{
print_html_style(theme = NULL)
}
\arguments{
\item{theme}{Theme overrides, see \code{theme} argument of \code{\link[=write_svg]{write_svg()}}.
\code{NULL} is the current default theme.}
}
\value{
Character vector of HTML/CSS lines that you can include in an
HTML file.
}
\description{
CSS for a theme
}
