% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_cast}
\alias{read_cast}
\title{Import an asciicast from an asciinema file}
\usage{
read_cast(json)
}
\arguments{
\item{json}{JSON asciinema file, version 2:
\url{https://github.com/asciinema/asciinema/blob/master/doc/asciicast-v2.md}.}
}
\value{
\code{asciicast} object.
}
\description{
Import an asciicast from an asciinema file
}
\seealso{
Other asciicast functions: \code{\link{play}},
  \code{\link{record}}, \code{\link{write_json}},
  \code{\link{write_svg}}
}
\concept{asciicast functions}
