% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{write_html}
\alias{write_html}
\title{Create a HTML snapshot of an asciicast}
\usage{
write_html(
  cast,
  path,
  at = "end",
  omit_last_line = NULL,
  prefix = "",
  full_html = FALSE,
  theme = NULL
)
}
\arguments{
\item{cast}{\code{asciicast} object.}

\item{path}{Path to the HTML file to create.}

\item{at}{When to take the snapshot, defaults to the end of the cast
(\code{"end"}). Can also be a number, in seconds.}

\item{omit_last_line}{Whether to omit the last line of the cast. This
often just the prompt, and sometimes it is not worth showing.}

\item{prefix}{Prefix to add to the beginning to every line. E.g.
\verb{#> } is usually added to knitr output.}

\item{full_html}{Whether to output a full HTML file. Mainly for testing
purposes.}

\item{theme}{A theme name to use, or a a named list to override the
default theme (see \code{\link[=default_theme]{default_theme()}}).}
}
\description{
Create a HTML snapshot of an asciicast
}
