% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{write_svg}
\alias{write_svg}
\title{Create animated SVG from an asciicast}
\usage{
write_svg(cast, path, window = NULL, start_at = NULL, end_at = NULL,
  at = NULL, cursor = NULL, rows = NULL, cols = NULL,
  padding = NULL, padding_x = NULL, padding_y = NULL)
}
\arguments{
\item{cast}{\code{asciicast} object.}

\item{path}{Path to SVG file to create.}

\item{window}{Render with window decorations.}

\item{start_at}{Lower range of timeline to render in seconds.}

\item{end_at}{Upper range of timeline to render in seconds.}

\item{at}{Timestamp of single frame to render, in seconds.}

\item{cursor}{Enable cursor rendering.}

\item{rows}{Height in lines.}

\item{cols}{Width in columns.}

\item{padding}{Distance between text and image bounds.}

\item{padding_x}{Distance between text and image bounds on x axis.}

\item{padding_y}{Distance between text and image bounds on y axis.}
}
\description{
Create animated SVG from an asciicast
}
\seealso{
Other asciicast functions: \code{\link{asciicast-package}},
  \code{\link{play}}, \code{\link{read_cast}},
  \code{\link{record}}, \code{\link{write_json}}
}
\concept{asciicast functions}
