% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{asciicast_start_process}
\alias{asciicast_start_process}
\title{Start an asciicast background process}
\usage{
asciicast_start_process(
  startup = NULL,
  timeout = 10,
  record_env = NULL,
  interactive = TRUE
)
}
\arguments{
\item{startup}{Quoted language object to run in the subprocess before
starting the recording.}

\item{timeout}{Idle timeout, in seconds If the R subprocess running
the recording does not answer within this limit, it is killed and the
recording stops.}

\item{record_env}{Environment variables to set for the R subprocess.}

\item{interactive}{Whether to run R in interactive mode. Note that in
interactive mode R might ask for terminal input.}
}
\value{
The R process, a \link[processx:process]{processx::process} object.
}
\description{
This is for expert use, if you want to run multiple recordings in the
same process.
}
\examples{
# Use the same R process to record multiple casts
process <- asciicast_start_process()
script1 <- "a <- runif(10)\n"
script2 <- "a\n"
cast1 <- record(textConnection(script1), process = process)
cast2 <- record(textConnection(script2), process = process)
cast1
cast2
}
\seealso{
Other asciicast functions: 
\code{\link{asciicast-package}},
\code{\link{read_cast}()},
\code{\link{record}()},
\code{\link{write_json}()}
}
\concept{asciicast functions}
