% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{init_knitr_engine}
\alias{init_knitr_engine}
\title{Initialize the asciicast knitr engine}
\usage{
init_knitr_engine()
}
\description{
Call this function in your Rmd file, to enable creating asciinema
casts from code chunks.
}
