% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frames.R
\name{load_frames}
\alias{load_frames}
\title{Extract / create complete screen frames from an ascii cast}
\usage{
load_frames(cast)
}
\arguments{
\item{cast}{aciicast object}
}
\value{
\code{asciicast_frames} object.
}
\description{
Extract / create complete screen frames from an ascii cast
}
\keyword{internal}
