% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{clear_screen}
\alias{clear_screen}
\alias{pause}
\alias{merge_casts}
\title{Merge multiple ASCII casts into one}
\usage{
clear_screen()

pause(secs)

merge_casts(...)
}
\arguments{
\item{secs}{Number of seconds to wait.}

\item{...}{Ascii casts to merge or merge commands. Merge commands
provide a way to insert pause, clear the screen, etc., between casts.}
}
\value{
An \code{asciicast} object.
}
\description{
\code{pause()} inserts a pause of the specified seconds between the casts.
}
\details{
\code{clear_screen()} clears the screen between two casts.
}
