% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciicast.R
\name{write_svg}
\alias{write_svg}
\title{Create animated SVG from a cast}
\usage{
write_svg(cast, path)
}
\arguments{
\item{cast}{\code{asciicast} object.}

\item{path}{Path to SVG file to create.}
}
\description{
Create animated SVG from a cast
}
