% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{write_svg}
\alias{write_svg}
\title{Create animated SVG from an ascii cast}
\usage{
write_svg(cast, path, window = TRUE, from = NULL, to = NULL,
  at = NULL, cursor = TRUE, height = NULL, width = NULL,
  padding = NULL, padding_x = NULL, padding_y = NULL)
}
\arguments{
\item{cast}{\code{asciicast} object.}

\item{path}{Path to SVG file to create.}

\item{window}{Render with window decorations.}

\item{from}{Lower range of timeline to render in seconds.}

\item{to}{Upper range of timeline to render in seconds.}

\item{at}{Timestamp of frame to render in seconds.}

\item{cursor}{Enable cursor rendering.}

\item{height}{Height in lines.}

\item{width}{Width in columns.}

\item{padding}{Distance between text and image bounds.}

\item{padding_x}{Distance between text and image bounds on x axis.}

\item{padding_y}{Distance between text and image bounds on y axis.}
}
\description{
Create animated SVG from an ascii cast
}
\seealso{
Other asciicast functions: \code{\link{play}},
  \code{\link{read_cast}}, \code{\link{record}},
  \code{\link{write_json}}
}
\concept{asciicast functions}
